Moved kernel to dedicated folder.
* also distributed kernels to files.
